fix(update): Lazy-link SHA512 so unused Update sketches keep FLASH
Move SHA-512 into its own TU and bind it only from setSHA512(), matching
the HTTPUpdateChecksum sidecar pattern so --gc-sections can drop mbedtls/PSA SHA-512.
